rename logic --> operators
