+++ corehq/apps/userreports/operators.py
